requirements.txt
anyio==4.9.0
asgiref==3.8.1
awsebcli==3.24.1
bcrypt==4.3.0
blessed==1.21.0
boto3==1.38.36
botocore==1.38.36
CacheControl==0.14.3
cachetools==5.5.2
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.3
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
fabric==3.2.2
firebase-admin==6.9.0
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
msgpack==1.1.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==1.26.20
wcwidth==0.2.13
wrapt==1.17.2
